apply generic search for export container

+++ src/main/java/com/justanalytics/dto/ContainerDto.java
@@ -39,7 +39,7 @@ public class ContainerDto {
     @JsonProperty(value = "EquipmentType")
     private String equipmentType;
     @JsonProperty(value = "TEU")
-    private Integer teu;
+    private Float teu;
     @JsonProperty(value = "OperatorLineID")
     private String operatorLineId;
     @JsonProperty(value = "OperatorName")
