+++ src/main/java/com/justanalytics/dto/EmptyContainerDto.java
@@ -37,7 +37,7 @@ public class EmptyContainerDto {
     @JsonProperty(value = "EquipmentType")
     private String equipmentType;
     @JsonProperty(value = "TEU")
-    private Integer teu;
+    private Float teu;
     @JsonProperty(value = "OperatorLineID")
     private String operatorLineId;
     @JsonProperty(value = "OperatorName")
